requirements.txt
gunicorn==20.1.0
psycopg2-binary==2.9.1
WTForms==2.3.3
Flask_Bootstrap==3.3.7.1
Flask==1.1.2
Flask_SQLAlchemy==2.5.1
Flask_WTF==0.15.1